# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Projecto challengeWyleex: Documentación de la API de lectores y de noticias",
        "description": "Listado de URI's: Esta documentación provee información necesaria para la utilización correcta de los End-Points",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://challengewyleex.darmsportfolio.xyz"
        },
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/login": {
            "post": {
                "tags": [
                    "Auth Login: Obtener token"
                ],
                "summary": "Login",
                "description": "Login",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "admin@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "admin123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vMTI3LhAuMC4xOjgwMDAvYXBpL3JlZnJlc2giLCJpYXQiOjE2ODA2MzUyNTMsImV4cCI6MTY4MTI0MDA5MCwibmJmIjoxNjgwNjM1MjkwLCJqdGkiOiJIR1o5M1JFbU1wc2hJTUt3Iiwic3ViIjoiMSIsInBydiI6IjIzYmQ1Yzg5NDlmNjAwYWRiMzll2zAxYzQwMDg3MmRiN2E1OTc2ZjcifQ.B8ogV3E5tMPfII1-vTvzqpXO_pt73II0koXkb63jEVQ"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "type": "number",
                                            "example": "604800"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/me": {
            "get": {
                "tags": [
                    "Auth Me: Obtener información del usuario logueado"
                ],
                "summary": "Información usuario logueado",
                "description": "Retorna los datos del usuario autentificado",
                "operationId": "9d9bc556f927575aa492a88c28fe5908",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "john@mail.com"
                                        },
                                        "email_verified_at": {
                                            "type": "string|null",
                                            "example": "null"
                                        },
                                        "role": {
                                            "type": "string",
                                            "example": "ADMIN"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-04-03T09:23:30.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-04-03T09:23:30.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth Logout: Cerrar Sesión"
                ],
                "summary": "Logout",
                "description": "Logout",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully logged out"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/refresh": {
            "post": {
                "tags": [
                    "Auth Refresh: Actualizar Token"
                ],
                "summary": "Refresh",
                "description": "Refresh",
                "operationId": "2c27d311a2876560fcb5a754b556cdc5",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vMTI3LhAuMC4xOjgwMDAvYXBpL3JlZnJlc2giLCJpYXQiOjE2ODA2MzUyNTMsImV4cCI6MTY4MTI0MDA5MCwibmJmIjoxNjgwNjM1MjkwLCJqdGkiOiJIR1o5M1JFbU1wc2hJTUt3Iiwic3ViIjoiMSIsInBydiI6IjIzYmQ1Yzg5NDlmNjAwYWRiMzll2zAxYzQwMDg3MmRiN2E1OTc2ZjcifQ.B8ogV3E5tMPfII1-vTvzqpXO_pt73II0koXkb63jEVQ"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "type": "number",
                                            "example": "604800"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}